split Login-page elements into 2 components
